python_requirements.txt
flask
irckit
gevent
redis
